services/data_proxy/requirements.txt
flask>=3.0.0
influxdb-client
paho-mqtt
